Drop cn-cbor dependency from CBOR cert operations
The CBOR cert operations only need to generate, not parse, CBOR which
has a fairly simple encoding. Move from using cn-cbor to directly
streaming to an output buffer.

Dropping cn-cbor also removes the dependency on allocation, which could
have been a problem in some environments, and removes the related class
of allocation errors.

The CBOR writing function are exposed to allow clients to make use of
them, for example to construct an encoded descriptor field.

According to pw_bloat, the library grows a little in size but the
consumer of the library shrinks to a size similar to that of the
template version.

Change-Id: I6281b166029fa79a67622112994d2b98ae5d9304
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/37740
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>
Reviewed-by: Darren Krahn <dkrahn@google.com>