Expose signing operations in DiceOps
The CBOR certificate generator previously had a dependency on boringssl
that meant it couldn't be reused with a different crypto implementation.
Remove the dependency by adding the needed operations to DiceOps so the
client can provide a suitable implementation. There are boringssl
implementations provided that can be used where boringssl is available.

The other certificate generators are tied to a crypto implementation so
don't need these operations to be provided.

Change-Id: Iee27061444da48c6dd2ef4cf91676b45fb227c7c
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/40021
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>
Reviewed-by: Darren Krahn <dkrahn@google.com>