Rename some private keys to private key seeds
The value derived by the profile is not necessarily used directly as the
private key but as a seed from which the private key is
deterministically generated. Mark this distinction by naming it as a
private key seed rather than key.

Change-Id: I09094c6177d999e5a95e1d825ddb3f56d1bbe24f
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/40741
Reviewed-by: Darren Krahn <dkrahn@google.com>
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>