Evolve the CBOR writer API
Treat struct CborOut as opaque in the API and add accessor functions for
the useful properties.

Remove error handling from each of the writing functions and instead
check for errors explicity on the struct CborOut. This encourages code
that optimisitically attempts to write everything to the buffer, the
common case, and only checks for success once complete. The space that
would have been neededi to encode everything is always captured to allow
for measurement. Code size also reduces a little as a result.

Change-Id: Idb35719087cb23eb6aafba1d11597c1b57f4f437
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/46140
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>
Reviewed-by: Darren Krahn <dkrahn@google.com>