Add CBOR reading utilities
The reading is very simplistic, parsing individual tokens. The larger
structure of the data would be interpreted by the client. There is a
function to skip over a more complex items but it doesn't extract any of
the data and just moves past it in the input buffer. Skipping supports a
limited depth of nesting and fails if the nesting goes too deep.

A fuzzer is included. As are a set of tests that no doubt could be
expanded upon.

Change-Id: I251e34aea69eba18a81edc42ab34a54a6a04b45b
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/65401
Reviewed-by: Darren Krahn <dkrahn@google.com>
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>