Add BCC functions
The Boot Certificate Chain (BCC) is an application of the Open Profile
for DICE that is used by Android. The BCC conforms to the Open Profile
for DICE standard and goes on to specify more strictly details such as
the configuration descriptor.

There is also an example of a data format used to pass the BCC between
boot stages. This BCC handover format encodes the CDIs as well as the
BCC.

Change-Id: Ib32bb0fba317cf5320d60e8479eb0484a6f90c71
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/65402
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>
Reviewed-by: Darren Krahn <dkrahn@google.com>