Roll boringssl
This CL also addresses an X.509 encoding quirk where a boolean was
previously encoded as 0x01 rather than 0xFF. This is not incorrect as
much as unconventional. The latest version of boringssl corrects this
and the known test values and templates here follow suit.

Test: Build, unit tests
Change-Id: I7ec778457fbfc16a5845d98a057b8d39b6bcc1c6
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/75361
Reviewed-by: Andrew Scull <ascull@google.com>
Commit-Queue: Darren Krahn <dkrahn@google.com>