Update DiceMainFlow to compute CDI certificate only if requested.
In the existing code, DiceMainFlow computes i) the two CDIs and
ii) the CDI certificate. However, in certain cases, we need
to compute only the two CDIs for the next stage.
For e.g. due to privacy reasons, certain parent stages might not
pass the CDI certificate to the next stage. In such cases,
the caller can signal this by passing zero/null values to the
certificate parameters.

Tests: Added a test to verify that passing null/zero values to
       DiceMainFlow doesn't cause panic.

Change-Id: I1b5c4939b82f79bf72c3f22d2d9d5a40c944a6cb
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/87560
Reviewed-by: Andrew Scull <ascull@google.com>
Commit-Queue: Hasini Gunasinghe <hasinitg@google.com>