Add BccHandoverParse
To help with accessing the fields of the handover, add a function to
parse the top-level CBOR map. Do this by extracting and exporting the
existing logic. Users of the library need to do this parsing e.g. to get
access to the sealing CDI in order to derive a key for sealing data.

Change-Id: If8b6ffa16f50c2872486d22de3ea05cf489ecb55
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/93561
Commit-Queue: Andrew Scull <ascull@google.com>
Pigweed-Auto-Submit: Andrew Scull <ascull@google.com>
Reviewed-by: Darren Krahn <dkrahn@google.com>
Reviewed-by: Hasini Gunasinghe <hasinitg@google.com>