Add functions to read and write CBOR tags
The functions to read and write tags do not perform any tag
validation; this step is left to callers. CborReadSkip treats
tags as a one-element array in order to correcly skip content
with nested tags. This change also modifies the test suite and
fuzzers to support tags.

Change-Id: Ic86b03739e1aaf06a4f0e47ad32fe63ed78cccbe
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/99900
Reviewed-by: Andrew Scull <ascull@google.com>
Commit-Queue: Andrew Scull <ascull@google.com>