Return required size if buffer is too small
The documentation for DiceMainFlow() says that the required buffer size
is returned if the provided buffer was too small. This stopped happening
in the CBOR implementation during the migration away from cn-cbor, so
fix that.

As the buffer is used both for the TBS data and the final result, care
is taken to find the size of buffer that is required for the whole
function to complete.

There are a number of APIs that follow a similar argument pattern so
also fix those to return the required size when the buffer is too small.

Change-Id: Ia5b01e36216761a58972a1afc8f747182470a631
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/67760
Reviewed-by: Darren Krahn <dkrahn@google.com>
Commit-Queue: Andrew Scull <ascull@google.com>