Remove description of RKP VM marker from Android profile docs
We hope to remove the RKP VM marker in the future, but if we
document it here, it will be difficult to remove. Favor documenting this marker in the Android HAL docs so we  can version it alongside of
Android. This will allow us to remove it from the docs once it's no
longer needed by Android.

Change-Id: I6915efc152fbfac6f000b28c2ca22341727139c8
Reviewed-on: https://pigweed-review.googlesource.com/c/open-dice/+/183159
Reviewed-by: Alan Stokes <alanstokes@google.com>
Reviewed-by: Andrew Scull <ascull@google.com>
Commit-Queue: Seth Moore <sethmo@google.com>